feat: extract event listener logic into a separate module
